requirements.txt
alembic==1.0.7
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
Click==7.0
cryptography==2.5
Flask==1.0.2
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.1
psycopg2==2.7.7
pycparser==2.19
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
python-telegram-bot==11.1.0
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.18
urllib3==1.24.1
Werkzeug==0.14.1
xmltodict==0.12.0
